DRY up issue create/edit handler tests
